[IFRT] Add ArraySpec
`ArraySpec` holds static properties of `Array`. This provides a common and
convenient way of describing an expected or requested static properties of an
input/output array of an operation.

This does not change IFRT Proxy's proto messages as they reflect the IFRT API,
most of which does not use `ArraySpec` yet.

PiperOrigin-RevId: 625023438